+++ qase-java-commons/src/main/java/io/qase/commons/config/ApiConfig.java
@@ -3,4 +3,5 @@ package io.qase.commons.config;
 public class ApiConfig {
     public String host = "qase.io";
     public String token = "";
+    public int timeoutSeconds = 0; // APIR-02: 0 = use OkHttp default (10s)
 }